[CodeCompletion] Move CompletionContextFinder to its own file
