+++ lib/Sema/TypeCheckCodeCompletion.cpp
@@ -43,6 +43,7 @@
 #include "swift/Sema/IDETypeChecking.h"
 #include "swift/Sema/CodeCompletionTypeChecking.h"
 #include "swift/Sema/ConstraintSystem.h"
+#include "swift/Sema/CompletionContextFinder.h"
 #include "swift/Strings.h"
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/ADT/PointerUnion.h"
@@ -511,221 +512,6 @@ TypeChecker::getTypeOfCompletionOperator(DeclContext *DC, Expr *LHS,
     llvm_unreachable("Invalid DeclRefKind for operator completion");
   }
 }
-
-namespace {
-
-class CompletionContextFinder : public ASTWalker {
-  enum class ContextKind {
-    FallbackExpression,
-    StringInterpolation,
-    SingleStmtClosure,
-    MultiStmtClosure,
-    ErrorExpression
-  };
-
-  struct Context {
-    ContextKind Kind;
-    Expr * E;
-  };
-
-  /// Stack of all "interesting" contexts up to code completion expression.
-  llvm::SmallVector<Context, 4> Contexts;
-
-  /// If we are completing inside an expression, the \c CodeCompletionExpr that
-  /// represents the code completion token.
-
-  /// The AST node that represents the code completion token, either as a
-  /// \c CodeCompletionExpr or a \c KeyPathExpr which contains a code completion
-  /// component.
-  llvm::PointerUnion<CodeCompletionExpr *, const KeyPathExpr *> CompletionNode;
-
-  Expr *InitialExpr = nullptr;
-  DeclContext *InitialDC;
-
-public:
-  /// Finder for completion contexts within the provided initial expression.
-  CompletionContextFinder(Expr *initialExpr, DeclContext *DC)
-      : InitialExpr(initialExpr), InitialDC(DC) {
-    assert(DC);
-    initialExpr->walk(*this);
-  };
-
-  /// Finder for completion contexts within the outermost non-closure context of
-  /// the code completion expression's direct context.
-  CompletionContextFinder(DeclContext *completionDC): InitialDC(completionDC) {
-    while (auto *ACE = dyn_cast<AbstractClosureExpr>(InitialDC))
-      InitialDC = ACE->getParent();
-    InitialDC->walkContext(*this);
-  }
-
-  std::pair<bool, Expr *> walkToExprPre(Expr *E) override {
-    if (auto *closure = dyn_cast<ClosureExpr>(E)) {
-      Contexts.push_back({closure->hasSingleExpressionBody()
-                            ? ContextKind::SingleStmtClosure
-                            : ContextKind::MultiStmtClosure,
-                          closure});
-    }
-
-    if (isa<InterpolatedStringLiteralExpr>(E)) {
-      Contexts.push_back({ContextKind::StringInterpolation, E});
-    }
-
-    if (isa<ApplyExpr>(E) || isa<SequenceExpr>(E)) {
-      Contexts.push_back({ContextKind::FallbackExpression, E});
-    }
-
-    if (auto *Error = dyn_cast<ErrorExpr>(E)) {
-      Contexts.push_back({ContextKind::ErrorExpression, E});
-      if (auto *OrigExpr = Error->getOriginalExpr()) {
-        OrigExpr->walk(*this);
-        if (hasCompletionExpr())
-          return std::make_pair(false, nullptr);
-      }
-    }
-
-    if (auto *CCE = dyn_cast<CodeCompletionExpr>(E)) {
-      CompletionNode = CCE;
-      return std::make_pair(false, nullptr);
-    }
-    if (auto *KeyPath = dyn_cast<KeyPathExpr>(E)) {
-      for (auto &component : KeyPath->getComponents()) {
-        if (component.getKind() ==
-            KeyPathExpr::Component::Kind::CodeCompletion) {
-          CompletionNode = KeyPath;
-          return std::make_pair(false, nullptr);
-        }
-      }
-      // Code completion in key paths is modelled by a code completion component
-      // Don't walk the key path's parsed expressions.
-      return std::make_pair(false, E);
-    }
-
-    return std::make_pair(true, E);
-  }
-
-  Expr *walkToExprPost(Expr *E) override {
-    if (isa<ClosureExpr>(E) || isa<InterpolatedStringLiteralExpr>(E) ||
-        isa<ApplyExpr>(E) || isa<SequenceExpr>(E) || isa<ErrorExpr>(E)) {
-      assert(Contexts.back().E == E);
-      Contexts.pop_back();
-    }
-    return E;
-  }
-
-  /// Check whether code completion expression is located inside of a
-  /// multi-statement closure.
-  bool locatedInMultiStmtClosure() const {
-    return hasContext(ContextKind::MultiStmtClosure);
-  }
-
-  bool locatedInStringIterpolation() const {
-    return hasContext(ContextKind::StringInterpolation);
-  }
-
-  bool hasCompletionExpr() const {
-    return CompletionNode.dyn_cast<CodeCompletionExpr *>() != nullptr;
-  }
-
-  CodeCompletionExpr *getCompletionExpr() const {
-    assert(hasCompletionExpr());
-    return CompletionNode.get<CodeCompletionExpr *>();
-  }
-
-  bool hasCompletionKeyPathComponent() const {
-    return CompletionNode.dyn_cast<const KeyPathExpr *>() != nullptr;
-  }
-
-  /// If we are completing in a key path, returns the \c KeyPath that contains
-  /// the code completion component.
-  const KeyPathExpr *getKeyPathContainingCompletionComponent() const {
-    assert(hasCompletionKeyPathComponent());
-    return CompletionNode.get<const KeyPathExpr *>();
-  }
-
-  /// If we are completing in a key path, returns the index at which the key
-  /// path has the code completion component.
-  size_t getKeyPathCompletionComponentIndex() const {
-    assert(hasCompletionKeyPathComponent());
-    size_t ComponentIndex = 0;
-    auto Components =
-        getKeyPathContainingCompletionComponent()->getComponents();
-    for (auto &Component : Components) {
-      if (Component.getKind() == KeyPathExpr::Component::Kind::CodeCompletion) {
-        break;
-      } else {
-        ComponentIndex++;
-      }
-    }
-    assert(ComponentIndex < Components.size() &&
-           "No completion component in the key path?");
-    return ComponentIndex;
-  }
-
-  struct Fallback {
-    Expr *E; ///< The fallback expression.
-    DeclContext *DC; ///< The fallback expression's decl context.
-    bool SeparatePrecheck; ///< True if the fallback may require prechecking.
-  };
-
-  /// As a fallback sometimes its useful to not only type-check
-  /// code completion expression directly but instead add some
-  /// of the enclosing context e.g. when completion is an argument
-  /// to a call.
-  Optional<Fallback> getFallbackCompletionExpr() const {
-    if (!hasCompletionExpr()) {
-      // Creating a fallback expression only makes sense if we are completing in
-      // an expression, not when we're completing in a key path.
-      return None;
-    }
-
-    Optional<Fallback> fallback;
-    bool separatePrecheck = false;
-    DeclContext *fallbackDC = InitialDC;
-
-    // Find the outermost fallback expression within the innermost error
-    // expression or multi-statement closure, keeping track of its decl context.
-    for (auto context: Contexts) {
-      switch (context.Kind) {
-      case ContextKind::StringInterpolation:
-        LLVM_FALLTHROUGH;
-      case ContextKind::FallbackExpression:
-        if (!fallback && context.E != InitialExpr)
-          fallback = Fallback{context.E, fallbackDC, separatePrecheck};
-        continue;
-
-      case ContextKind::SingleStmtClosure:
-        if (!fallback && context.E != InitialExpr)
-          fallback = Fallback{context.E, fallbackDC, separatePrecheck};
-        fallbackDC = cast<AbstractClosureExpr>(context.E);
-        continue;
-
-      case ContextKind::MultiStmtClosure:
-        fallbackDC = cast<AbstractClosureExpr>(context.E);
-        LLVM_FALLTHROUGH;
-      case ContextKind::ErrorExpression:;
-        fallback = None;
-        separatePrecheck = true;
-        continue;
-      }
-    }
-
-    if (fallback)
-      return fallback;
-
-    if (getCompletionExpr()->getBase() && getCompletionExpr() != InitialExpr)
-      return Fallback{getCompletionExpr(), fallbackDC, separatePrecheck};
-    return None;
-  }
-
-private:
-  bool hasContext(ContextKind kind) const {
-    return llvm::find_if(Contexts, [&kind](const Context &currContext) {
-             return currContext.Kind == kind;
-           }) != Contexts.end();
-  }
-};
-
-} // end namespace
 
 /// Remove any solutions from the provided vector that both require fixes and
 /// have a score worse than the best.
